--- src/components/BreadcrumbsItem/index.ts
@@ -1,1 +1,0 @@
-export { BreadcrumbsItem as default } from './BreadcrumbsItem';
